+++ TEST/prj.conf
@@ -74,3 +74,6 @@ CONFIG_USE_SEGGER_RTT=y
 CONFIG_RTT_CONSOLE=y
 CONFIG_UART_CONSOLE=n
 #-----------------------------------------------------------------------------
+
+
+CONFIG_BT_DEVICE_NAME_DYNAMIC=y
